\name{calc_vase}
\alias{calc_vase}
\title{Calculations for vase plots.}
\usage{
  calc_vase(x, ..., names = NULL, bw = NULL)
}
\arguments{
  \item{x}{continuous variable}

  \item{...}{other variables}

  \item{names}{labels}

  \item{bw}{bandwidth for kernel density}
}
\value{
  list of numeric descriptions of body, whiskers, outliers
  and the median
}
\description{
  Calculations for vase plots.
}
\author{
  Lisa Stryjewski
}

