% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{ggboxplots}
\alias{ggboxplots}
\alias{ggboxplots-package}
\title{A package for creating common variations of boxplots}
\description{
This package brings a set of variations of boxplots to the \code{ggplot2}.framework.
Implemented are vase plots, violin plots, and high density
region plots.
See \code{\link{ggvase}, \link{ggviolin}} and \code{\link{gghdr}}
}
\examples{
require(ggplot2)
ggviolin(diamonds, cut, price, bandwidth=300, fill=cut) + scale_fill_brewer(palette="Set1")
gghdr(diamonds, cut, price, probs=c(50,25,12.5, 6.25), fill=cut) +
  scale_fill_brewer(palette="Set1") +
  scale_colour_brewer(palette="Set1")
ggvase(diamonds, cut, price, bandwidth=300, fill=cut) + scale_fill_brewer(palette="Set1")
}
\references{
benjamini hintze hyndman
}

