\docType{package}
\name{package-ggboxplots}
\alias{package-ggboxplots}
\alias{package-ggboxplots-package}
\title{A package for creating common variations of boxplots}
\description{
  Boxplot variations, such as vase plots, violin plots, and
  high density region plots. See \cite{\link ggvase,
  ggviolin} and \cite{\link{gghdr}} Eventually these
  functions are supposed to go into the
  \code{\link{GGally}} package.
}

