% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stat-hdr.r
\name{stat_hdr}
\alias{stat_hdr}
\title{High Density Region plots.}
\usage{
stat_hdr(mapping = NULL, data = NULL, geom = "hdr", position = "dodge",
  show.legend = NA, inherit.aes = TRUE, width = 0.9, probs = c(0.99,
  0.95, 0.5), ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), is combined with the default mapping at the top level of the
plot. You only need to supply \code{mapping} if there isn't a mapping
defined for the plot.}

\item{data}{A data frame. If specified, overrides the default data frame
defined at the top level of the plot.}

\item{geom}{Use to override the default connection between
\code{geom_density} and \code{stat_density}.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{...}{other arguments passed on to \code{\link{layer}}. There are
  three types of arguments you can use here:

  \itemize{
  \item Aesthetics: to set an aesthetic to a fixed value, like
     \code{color = "red"} or \code{size = 3}.
  \item Other arguments to the layer, for example you override the
    default \code{stat} associated with the layer.
  \item Other arguments passed on to the stat.
  }}

\item{scale}{if "area" (default), all vases have the same area (before trimming
the tails). If "count", areas are scaled proportionally to the number of
observations. If "width", all vases have the same maximum width.}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with
   a warning. If \code{TRUE} silently removes missing values.}
}
\value{
A data frame with additional columns:
  \item{density}{density estimate}
  \item{scaled}{density estimate, scaled to maximum of 1}
  \item{count}{density * number of points - not sure about usefulness for vase plots}
  \item{vasewidth}{density scaled for the vase plot, according to area, counts
                     or to a constant maximum width}
  \item{n}{number of points}
  \item{width}{width of vase bounding box}
}
\description{
High density region (HDR) boxplots are a variation of boxplots.  HDR plots are based on a density estimate of the marginal distribution.
Cutoff values for the probability are defined in the parameter \code{probs} to define regions. In a uni-modal situation, the HDR plots with probability \code{probs=0.25}
show the boxes of a regular boxplot. The code is based on the \code{hdrcde} package by \cite{hyndman}, who also introduced the plots.
}
\examples{
# See geom_hdr for examples
# Also see stat_density for similar examples with data along x axis
}
\seealso{
\code{\link{geom_hdr}} for examples, and \code{\link{geom_violin}}
  for examples with data along the x axis.
}

