% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdr.r
\name{gghdr}
\alias{gghdr}
\title{High Density Region plots.}
\usage{
gghdr(data, x, y, probs = c(90, 50, 25), fill = "grey50", ...)
}
\arguments{
\item{data}{dataset}

\item{x}{factor variable}

\item{y}{values}

\item{probs}{vector of probabilities as cut-offs for the density regions}

\item{fill}{fill color of highest density regions. Either a color description (name or hex string), or grouping variable.}

\item{...}{other parameters passed on to the ggplot2 framework}
}
\value{
ggplot2 layer of highest density region boxplots.
}
\description{
High density region (HDR) boxplots are a variation of boxplots.  HDR plots are based on a density estimate of the marginal distribution.
Cutoff values for the probability are defined in the parameter \code{probs} to define regions. In a uni-modal situation, the HDR plots with probability \code{probs=0.25}
show the boxes of a regular boxplot. The code is based on the \code{hdrcde} package by \cite{hyndman}, who also introduced the plots.
}
\examples{
require(ggplot2)
data(diamonds)
gghdr(diamonds, color, price)
gghdr(diamonds, cut, price, probs=c(50,25,12.5, 6.25), fill=cut) +
  scale_fill_brewer(palette="Set1") +
  scale_colour_brewer(palette="Set1")
}
\author{
Heike Hofmann
}
\references{
hyndman
}

