\docType{package}
\name{package-ggboxplots}
\alias{package-ggboxplots}
\alias{package-ggboxplots-package}
\title{A package for creating common variations of boxplots}
\description{
  Boxplot variations, such as vase plots, violin plots,
  others. Eventually these functions are supposed to go
  into the \code{\link{GGally}} package.
}

