\docType{package}
\name{package-ggboxplots}
\alias{package-ggboxplots}
\alias{package-ggboxplots-package}
\title{A package for creating common variations of boxplots}
\description{
  Boxplot variations, such as vase plots, violin plots, and
  high density region plots. See \code{\link{ggvase},
  \code{ggviolin}} and \code{\link{gghdr}}
}

